refactor(vue-wrapper): remove vue-wrapper and its testing Remove vue-wrapper and extract vue responsibilities to independent files. EX-1885
